requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.12.3
numpy==2.2.6
langchain==1.2.9
langchain-openai==1.1.7
langchain-community==0.4.1
langgraph==1.0.8
skyfield==1.54
python-dotenv==1.1.0
pillow==11.3.0
tensorflow-cpu==2.20.0